Added bind host / port
On some systems you need to specify which network adapter / port you
want the outbound connection to be performed on. This is especially
common in cloud hosting environment with strict firewalls.

This commit provides the ability to specify a "bindaddress" and a
"bindport" parameters in the constructor function. These parameters
can just as easily be ignored and their default values of "0.0.0.0"
and "0" respectively will be used.